requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.1
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pypdfium2==4.30.1
pytest==8.3.5
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
